+++ source-code/sdk-js/src/adapter-sveltekit/ast-transformation/transforms/hooks.server.js.test.ts
@@ -44,7 +44,7 @@ describe("transformHooksServerJs", () => {
 	})
 })
 
-// NOTES
+// NOTES @stepan
 // Only one function for everything (transformHooksServerJs)
 // Capabilities:
 // 1. merge required imports with current imports